+++ src/tenant/admin.py
@@ -1,6 +1,3 @@
-import operator
-from functools import reduce
-
 from django import forms
 from django.conf import settings
 from django.contrib import admin, messages
@@ -10,26 +7,28 @@ from django.core.exceptions import PermissionDenied
 from django.contrib.sites.models import Site
 from django.template.response import TemplateResponse
 from django.db import connection, transaction
-from django.db.models import Q, F
 from django.utils.translation import ngettext
-from django.utils.text import smart_split, unescape_string_literal
 from django.contrib.admin.utils import unquote
 from django.contrib.admin.exceptions import DisallowedModelAdminToField
 from django.contrib.admin.options import TO_FIELD_VAR, IS_POPUP_VAR
 from django.utils.translation import gettext as _
+from django.utils.html import format_html
+from django.utils.formats import date_format
 from django.urls import reverse
 
 from allauth.socialaccount.models import SocialApp
 from allauth.account.models import EmailAddress
 from allauth.account.utils import user_email
-from django_tenants.utils import tenant_context, schema_context, get_public_schema_name
+from django_tenants.utils import get_public_schema_name, tenant_context
 
 from bytedeck_summernote.widgets import ByteDeckSummernoteSafeWidget
 from siteconfig.models import SiteConfig
-from tenant.models import Tenant, TenantDomain
-from tenant.forms import TenantBaseForm
-from tenant.utils import generate_schema_name
-from tenant.tasks import send_email_message
+
+from .models import Tenant, TenantDomain
+from .forms import TenantBaseForm
+from .utils import generate_schema_name
+from .tasks import send_email_message
+
 
 User = get_user_model()
 
@@ -124,15 +123,15 @@ class DeleteConfirmationForm(forms.Form):
 class TenantAdmin(PublicSchemaOnlyAdminAccessMixin, admin.ModelAdmin):
     list_display = (
         'schema_name', 'owner_full_name_text', 'owner_full_name_deprecated',
-        'owner_email_text', 'owner_email_deprecated',
+        'owner_email_text', 'owner_email_verified_boolean', 'owner_email_deprecated',
         'last_staff_login', 'google_signon_enabled',
-        'paid_until', 'trial_end_date',
+        'paid_until_text', 'trial_end_date_text',
         'max_active_users', 'active_user_count', 'total_user_count',
         'max_quests', 'quest_count',
     )
     list_filter = ('paid_until', 'trial_end_date', 'active_user_count', 'last_staff_login')
     # DEPRECATED: these fields ("owner_full_name" and "owner_email") will be removed in a future update
-    search_fields = ['schema_name', 'owner_full_name', 'owner_email']
+    search_fields = ['schema_name', 'owner_full_name', 'owner_full_name_cached', 'owner_email', 'owner_email_cached']
 
     form = TenantAdminForm
     inlines = (TenantDomainInline, )
@@ -140,22 +139,12 @@ class TenantAdmin(PublicSchemaOnlyAdminAccessMixin, admin.ModelAdmin):
     delete_selected_confirmation_template = 'admin/tenant/tenant/delete_selected_confirmation.html'
     delete_confirmation_template = 'admin/tenant/tenant/delete_confirmation.html'
 
-    actions = ['message_selected', 'enable_google_signin', 'disable_google_signin']
+    actions = ['message_unverified', 'message_verified', 'enable_google_signin', 'disable_google_signin']
 
     @admin.display(description="owner full name")
     def owner_full_name_text(self, obj):
-        """
-        Returns full name (or username) from SiteConfig().deck_owner object.
-        """
-        if obj.schema_name == get_public_schema_name():
-            return  # skip public schema
-
-        full_name_or_username = None
-        with tenant_context(obj):
-            owner = SiteConfig.get().deck_owner
-            # get the full name of the user, or if none is supplied will return the username
-            full_name_or_username = owner.get_full_name() or owner.username
-        return full_name_or_username
+        return obj.owner_full_name_cached
+    owner_full_name_text.admin_order_field = "owner_full_name_cached"
 
     @admin.display(description="owner full name (DEPRECATED)")
     def owner_full_name_deprecated(self, obj):
@@ -165,21 +154,25 @@ class TenantAdmin(PublicSchemaOnlyAdminAccessMixin, admin.ModelAdmin):
 
     @admin.display(description="owner email")
     def owner_email_text(self, obj):
+        return obj.owner_email_cached
+    owner_email_text.admin_order_field = "owner_email_cached"
+
+    @admin.display(description="email verified", boolean=True)
+    def owner_email_verified_boolean(self, obj):
         """
-        Returns email address (primary and verified) from SiteConfig().deck_owner object.
+        Returns `True` (green), if at least one verified email address was found, otherwise `False`.
         """
         if obj.schema_name == get_public_schema_name():
             return  # skip public schema
 
-        email = None
         with tenant_context(obj):
             owner = SiteConfig.get().deck_owner
             # get the email address, but only primary and verified
             for primary_email_address in EmailAddress.objects.filter(user=owner, primary=True, verified=True):
                 # make sure it's primary email for real
                 if primary_email_address.email == user_email(owner):
-                    email = owner.email
-        return email
+                    return True
+        return False
 
     @admin.display(description="owner email (DEPRECATED)")
     def owner_email_deprecated(self, obj):
@@ -187,66 +180,27 @@ class TenantAdmin(PublicSchemaOnlyAdminAccessMixin, admin.ModelAdmin):
         return obj.owner_email
     owner_email_deprecated.admin_order_field = "owner_email"
 
-    def get_search_results(self, request, queryset, search_term):
-        """
-        The `get_search_results` method modifies the list of objects displayed into those
-        that match the provided search term. It accepts the request, a queryset that applies
-        the current filters, and the user-provided search term. It returns a tuple containing a
-        queryset modified to implement the search, and a boolean indicating if the results
-        may contain duplicates.
+    @admin.display(description="paid until", ordering="paid_until")
+    def paid_until_text(self, obj):
+        """Returns htmlized value of `paid_until` field"""
+        if not obj.paid_until:
+            return None
+        return format_html(
+            "<span data-date=\"{}\">{}</span>",
+            obj.paid_until,
+            date_format(obj.paid_until, use_l10n=True),
+        )
 
-        This method was overridden to perform search on custom fields.
-        """
-        # The default implementation searches the fields named in ModelAdmin.search_fields.
-        queryset, may_have_duplicates = super().get_search_results(request, queryset, search_term)
-
-        # search_term is what you input in admin site
-        if not search_term:
-            return queryset, False
-
-        # getting a list of all schemas, except the public schema name
-        schemas = list(Tenant.objects.all().values_list("schema_name", flat=True))
-        schemas.remove(get_public_schema_name())
-
-        # perform search on custom fields: "owner_full_name", "owner_email"
-        search_fields = ["deck_owner__first_name", "deck_owner__last_name", "deck_owner__username", "deck_owner__email"]
-
-        # you need to take a copy of the list and iterate over it first,
-        # or the iteration will fail with what may be unexpected results.
-        for schema in schemas[:]:
-            with schema_context(schema):
-                orm_lookups = ["%s__icontains" % str(search_field) for search_field in search_fields]
-
-                for bit in smart_split(search_term):  # ["spam", "egg"]
-                    if bit.startswith(('"', "'")) and bit[0] == bit[-1]:
-                        bit = unescape_string_literal(bit)
-                    or_queries = []
-                    for orm_lookup in orm_lookups:
-                        # get the email address, but only primary and verified
-                        if "deck_owner__email" in orm_lookup:
-                            or_queries.append(Q(**{
-                                "deck_owner__emailaddress__verified": True,
-                                "deck_owner__emailaddress__primary": True,
-                                # make sure it's primary email for real
-                                "deck_owner__emailaddress__email__exact": F("deck_owner__email"),
-                                orm_lookup: bit,
-                            }))
-                        else:
-                            or_queries.append(Q(**{orm_lookup: bit}))
-                    qs = SiteConfig.objects.filter(reduce(operator.or_, or_queries))
-
-                # if nothing was found within schema context
-                if not qs.exists():
-                    schemas.remove(schema)
-
-        # return original queryset, OR any new results we found by searching within schema context
-        if len(schemas):
-            queryset |= self.model.objects.filter(schema_name__in=schemas)
-
-        # we must establish if the queryset changes implemented by our search method
-        # may introduce duplicates into the results, and return True in the second
-        # element of the return value.
-        return queryset, True
+    @admin.display(description="trial end date", ordering="trial_end_date")
+    def trial_end_date_text(self, obj):
+        """Returns htmlized value of `trial_end_date` field"""
+        if not obj.trial_end_date:
+            return None
+        return format_html(
+            "<span data-date=\"{}\">{}</span>",
+            obj.trial_end_date,
+            date_format(obj.trial_end_date, use_l10n=True),
+        )
 
     def get_actions(self, request):
         """
@@ -263,6 +217,8 @@ class TenantAdmin(PublicSchemaOnlyAdminAccessMixin, admin.ModelAdmin):
     def get_queryset(self, request):
         qs = super().get_queryset(request)
         # Update cached fields
+        #
+        # FIX: reduce number of SQL queries by triggering this only when Tenant (and/or related) objects are changed
         for tenant in qs:
             if tenant.name != get_public_schema_name():
                 with tenant_context(tenant):
@@ -361,10 +317,19 @@ class TenantAdmin(PublicSchemaOnlyAdminAccessMixin, admin.ModelAdmin):
 
         return self.render_delete_form(request, context)
 
-    @admin.action(description="Send an email message to all owners for the selected tenant(s)")
-    def message_selected(self, request, queryset):
+    @admin.action(description="Send an email message to *all* owners for the selected tenant(s)")
+    def message_unverified(modeladmin, request, queryset):
+        """Send an email message to *all* owners for selected tenant(s)."""
+        return modeladmin.message_selected(request, queryset, verified=False)
+
+    @admin.action(description="Send an email message to *verified* only owners for the selected tenant(s)")
+    def message_verified(modeladmin, request, queryset):
+        """Send an email message to *verified* only owners for selected tenant(s)."""
+        return modeladmin.message_selected(request, queryset, verified=True)
+
+    def message_selected(modeladmin, request, queryset, verified=True):
         """
-        Send an email message to all owners for selected tenant(s).
+        Send an email message to either *all* or *verified* only owners for selected tenant(s).
 
         This action first displays an intermediate page which shows compose message form and
         a list of all recipients.
@@ -372,7 +337,7 @@ class TenantAdmin(PublicSchemaOnlyAdminAccessMixin, admin.ModelAdmin):
         Next, it send email message to all selected users and redirects back to the change list.
         """
         # get a list of selected tenant(s), excluding public schema
-        objects = self.model.objects.filter(
+        objects = modeladmin.model.objects.filter(
             pk__in=[str(x) for x in request.POST.getlist(helpers.ACTION_CHECKBOX_NAME)]
         ).exclude(schema_name=get_public_schema_name())
 
@@ -386,9 +351,11 @@ class TenantAdmin(PublicSchemaOnlyAdminAccessMixin, admin.ModelAdmin):
                 # get the full name of the user, or if none is supplied will return the username
                 full_name_or_username = owner.get_full_name() or owner.username
 
-                # get the email address, but only primary and verified
                 email = ""
-                for primary_email_address in EmailAddress.objects.filter(user=owner, primary=True, verified=True):
+                for primary_email_address in EmailAddress.objects.filter(user=owner):
+                    # get the email address, but only primary and verified
+                    if verified and not (primary_email_address.verified and primary_email_address.primary):
+                        continue
                     # make sure it's primary email for real
                     if primary_email_address.email == user_email(owner):
                         email = owner.email
@@ -399,7 +366,7 @@ class TenantAdmin(PublicSchemaOnlyAdminAccessMixin, admin.ModelAdmin):
         # Create a message informing the user that the recipients were not found
         # and return a redirect to the admin index page.
         if not len(recipient_list):
-            self.message_user(request, "No recipients found.", messages.WARNING)
+            modeladmin.message_user(request, "No recipients found.", messages.WARNING)
             # return None to display the change list page again.
             return None
 
@@ -423,7 +390,7 @@ class TenantAdmin(PublicSchemaOnlyAdminAccessMixin, admin.ModelAdmin):
                 )
 
                 # show alert that everything is cool
-                self.message_user(request, "%s users emailed successfully!" % len(recipient_list), messages.SUCCESS)
+                modeladmin.message_user(request, "%s users emailed successfully!" % len(recipient_list), messages.SUCCESS)
 
                 # return None to display the change list page again.
                 return None
@@ -437,23 +404,23 @@ class TenantAdmin(PublicSchemaOnlyAdminAccessMixin, admin.ModelAdmin):
         # AdminForm is not usually used directly by developers, but can be used by libraries that want to
         # extend the forms within the Django Admin.
         adminform = helpers.AdminForm(form, [(None, {"fields": form.base_fields})], {})
-        media = self.media + adminform.media
+        media = modeladmin.media + adminform.media
 
-        request.current_app = self.admin_site.name
+        request.current_app = modeladmin.admin_site.name
 
         # we need to create a template of intermediate page with form
         return TemplateResponse(
             request,
             "admin/tenant/tenant/message_selected_compose.html",
             context={
-                **self.admin_site.each_context(request),
+                **modeladmin.admin_site.each_context(request),
                 "title": "Write your message here",
                 "adminform": adminform,
                 "subtitle": None,
                 "recipient_list": recipient_list,
                 "queryset": objects,
                 # building proper breadcrumb in admin
-                "opts": self.model._meta,
+                "opts": modeladmin.model._meta,
                 "DEFAULT_FROM_EMAIL": settings.DEFAULT_FROM_EMAIL,
                 "action_checkbox_name": helpers.ACTION_CHECKBOX_NAME,
                 "media": media,